python/requirements.txt
fastapi==0.115.12
llama_cpp_python==0.3.9
outlines==1.0.2
pydantic==2.11.5